Sem-Ver: bugfix Fix the requires_asap decorator for python 3 by forcing the HTTP_AUTHORIZATION header into bytes before parsing it
More issues with string encodings. The requests library seems to force
HTTP headers to be strings even if you give it a bytes so to deal with
that potential ambiguity, we should just explicitly convert any
strings we get in the HTTP_AUTHORIZATION header into bytes before
attempting to parse it further.