Wrap request exceptions in public key retrievers
Problems retrieving public keys should show up as
PublicKeyRetrieverException.

Note that the asynchronous client raises a different exception, which
must be wrapped separately.

Also note that Python 2 and Python 3 handle raising exceptions very
differently, so we punt on the problem of `raise ..., ..., ...` vs.
`raise ... from ...` for now and just staple the original exception to
the new exception.

Sem-Ver: api-break