Raise PublicKeyRetrieverException uniformly
Instead of raising whatever exception comes out of the underlying
request library, use `PublicKeyRetrieverException` so that callers don't
have to know which retriever is being used.

If the original exception includes an HTTP status code, it should be
available in the `status_code` attribute. This should allow for more
uniform handling of bad requests and remote server errors.

If a caller cares about the underlying exception, it should be available
in the `original_exception` attribute.

Sem-Ver: api-break