Rework Django and Flask support
Previously, contributors introduced Django and Flask support via
differing implementations into the contrib/ package. The implementations
were similar (sometimes almost copied and pasted from one another) but
also diverged in terms of their functionality and feature sets.

This change's primary goal is to take the contrib/{django,flask} and
refactor them into a shared implementation with framework-specific
backends to abstract away the differences between the two. To
demonstrate this formalization of support for these frameworks, they
have also been moved from contrib into frameworks/{django,flask}.

In addition the following changes were made in this commit:

- Django and Flask decorators no longer require an issuer to be present
in BOTH the global setting and the decorator argument to be valid.
Decorator arguments now override the global setting.
- The proxied middleware seems like a one-off special case and remains
in contrib while a more general ASAP middleware which applies the
decorator logic to all views is included in the supported package.
- The @require_asap decorator has been renamed to @with_asap since you
can technically pass the required=False argument to make ASAP not
enforced.
- The @validate_asap decorator has been renamed to a more descriptive
@restrict_asap to symbolize that it applies additional restrictions onto
an already processed ASAP authenticated request.
- The 'subjects' arg to the @validate_asap decorator has been removed
since it appears to be unused and is inconsistent with the existing
@with_asap decorator.

Sem-Ver: api-break