Sem-Ver: feature Add WSGI middleware
Fix a broken test

Enhance tests for WSGI middleware

Fix extra line at the end of a file

Modify WSGI backend to check if request is None instead of falsey

Fix some failing tests

Fix another test

Don't initialize verifier in _process_asap_token if there is no token

Pass through the merged settings to backend.get_verifier.

Signed-off-by: David Black <dblack@atlassian.com>

When creating an instance of JWTAuthVerifier we should not provide a subject_should_match_issuer keyword argument if the settings value is None.

Signed-off-by: David Black <dblack@atlassian.com>

Add a test for when the django decorator is used to override the settings defined subject_does_need_to_match_issuer value.

Signed-off-by: David Black <dblack@atlassian.com>

Remove the unused SubjectIssuerMismatchError exception.

Signed-off-by: David Black <dblack@atlassian.com>

Restore the docstring for requires_asap.

Signed-off-by: David Black <dblack@atlassian.com>

Do not catch the InvalidTokenError as an exception variable.

Signed-off-by: David Black <dblack@atlassian.com>